+++ source/uim/cake/https/cookies/collection.d
@@ -68,7 +68,7 @@ class CookieCollection : IteratorAggregate, Countable {
     /**
      * Get the number of cookies in the collection.
      */
-    int count() {
+    size_t count() {
         return count(this.cookies);
     }
