+++ source/uim/cake/orm/query.d
@@ -823,7 +823,7 @@ class Query : DatabaseQuery : JsonSerializable, IQuery {
      * modified, and the count has already been performed the cached
      * value is returned
      */
-    int count() {
+    size_t count() {
         if (_resultsCount is null) {
             _resultsCount = _performCount();
         }
